+++ tests/bot.test.js
@@ -244,18 +244,6 @@ describe('mwn', async function () {
 		expect(bot.Date.localeData.daysShort[0]).to.equal('dim.');
 	});
 
-	it('deprecated smallcase class names still work', async function () {
-		await bot.Date.populateLocaleData('fr');
-		expect(bot.Date.localeData.months[0]).to.equal('janvier');
-
-		const title = bot.title.newFromText('caTeGoRy:living people');
-		expect(title.getNamespaceId()).to.equal(14);
-		expect(title.getPrefixedText()).to.equal('Category:Living people');
-
-		const user = new bot.user('SD0001');
-		expect(await user.contribs({ uclimit: 1 })).to.be.of.length(1);
-	});
-
 	//////////////////////////////////////////
 	// UNSUCCESSFUL                         //
 	//////////////////////////////////////////